Block the removal of the Realm primary external ids
Do not allow removing e-mails associated with the primary
external id of a Realm as it would result in effectively
removing the account id in that realm.

Removing accounts is not allowed in Gerrit and therefore
should not be allowed in Realms as well, especially when
that is done unconsciously while removing an e-mail.

This problem has been in the Gerrit code-base for years
but it was hidden by the front-end check done in the GWT UI
that was checking if the realm allowed or not that deletion.
The advent of PolyGerrit UI has changed how things work
and delegated the responsibility of making the decision to
the backend, which was broken.

The attempt to remove the e-mail associated with the primary
external id scheme of the realm will now result in a
ResourceConflictException, which is one of the error conditions
that should be already known to the PolyGerrit UI.

Bug: Issue 12755

+++ java/com/google/gerrit/server/restapi/account/DeleteEmail.java
@@ -95,6 +95,14 @@ public class DeleteEmail implements RestModifyView<AccountResource.Email, Input>
       throw new ResourceNotFoundException(email);
     }
 
+    if (realm.accountBelongsToRealm(extIds)) {
+      String errorMsg =
+          String.format(
+              "Removal of e-mail %s not allowed because it is associated with the main scheme of the realm",
+              email);
+      throw new ResourceConflictException(errorMsg);
+    }
+
     try {
       accountManager.unlink(
           user.getAccountId(), extIds.stream().map(ExternalId::key).collect(toSet()));
